refactor(db): split db::models tests to sibling file for 1000-line cap
models.rs grew to 1052 lines after the mutation-coverage tests added in
7926b27. Extract the test module to db/models_tests.rs via the same
`#[path] mod tests;` pattern used for db/mod.rs and api/routes.rs.
